Update PrimaryFile
Corrected a typo that incorrectly gave a result that appeared to be correct.  Now each density field uses its own colour density scale so they are easier to compare side by side. 

Produces a vector field of neutrino velocities however the results of this do not work yet, along with a scatter plot of neutrino positions that is so dense it appears to be uniform.